sample validations

+++ src/Domain/Entities/Order.cs
@@ -14,7 +14,7 @@ namespace Sanoma.Domain.Entities
 
 		public DateTime? SubmitDate { get; set; }
 
-		public double TotalAmount { get; set; }
+		public decimal TotalAmount { get; set; }
 
 		public bool WillCall { get; set; }
     }
